+++ src/entity/on_page/pages.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 pub struct OnPageDataApiPage {
     pub crawl_progress: Option<String>,
     pub crawl_status: Option<OnPageDataApiCrawlStatus>,
-    pub total_items_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub total_items_count: Option<i64>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<OnPageResourceHtml>>,
 }
