+++ src/entity/on_page/waterfall.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 pub struct OnPageDataApiWaterfall {
     pub crawl_progress: Option<String>,
     pub crawl_status: Option<OnPageDataApiCrawlStatus>,
-    pub items_count: Option<i32>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<OnPageDataApiWaterfallItem>>,
 }
