+++ src/entity/serp/google/event.rs
@@ -15,7 +15,7 @@ pub struct SerpApiGoogleEventAdvance {
     pub spell: Option<SerpApiGoogleOrganicTaskSpell>,
     pub refinement_chips: Option<SerpApiElementRefinementChips>,
     pub item_types: Option<Vec<String>>,
-    pub se_results_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub se_results_count: Option<i64>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<SerpApiElementEventItem>>,
 }
