+++ src/entity/serp/google/jobs.rs
@@ -16,8 +16,8 @@ pub struct SerpApiGoogleJob<T> {
     pub spell: Option<Value>,
     pub refinement_chips: Option<SerpApiElementRefinementChips>,
     pub item_types: Option<Vec<String>>,
-    pub se_results_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub se_results_count: Option<i64>,
+    pub items_count: Option<i64>,
     pub items: Option<Vec<T>>,
 }
