+++ src/entity/serp/google/organic.rs
@@ -21,8 +21,8 @@ pub struct SerpApiGoogleOrganicTaskRegular {
     pub spell: Option<SerpApiGoogleOrganicTaskSpell>,
     pub refinement_chips: Option<SerpApiElementRefinementChips>,
     pub item_types: Option<Vec<String>>,
-    pub se_results_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub se_results_count: Option<i64>,
+    pub items_count: Option<i64>,
     // pub items: Option<Vec<Value>>,
     pub items: Option<Vec<SerpApiGoogleOrganicItem>>,
 }
@@ -39,8 +39,8 @@ pub struct SerpApiGoogleOrganicTaskAdvanced {
     pub spell: Option<SerpApiGoogleOrganicTaskSpell>,
     pub refinement_chips: Option<SerpApiElementRefinementChips>,
     pub item_types: Option<Vec<String>>,
-    pub se_results_count: Option<i32>,
-    pub items_count: Option<i32>,
+    pub se_results_count: Option<i64>,
+    pub items_count: Option<i64>,
     // pub items: Option<Vec<Value>>,
     pub items: Option<Vec<SerpApiGoogleOrganicItem>>,
 }
